+++ tests/unit/Codeception/Util/LocatorTest.php
@@ -53,6 +53,15 @@ class LocatorTest extends PHPUnit_Framework_TestCase
         $this->assertFalse(Locator::isID('hello'));
     }
 
+    public function testContains()
+    {
+        $this->assertEquals('descendant-or-self::label[contains(., \'enter a name\')]', Locator::contains('label', 'enter a name'));
+        $this->assertEquals('descendant-or-self::label[@id = \'user\'][contains(., \'enter a name\')]', Locator::contains('label#user', 'enter a name'));
+        $this->assertEquals('//label[@for="name"][contains(., \'enter a name\')]', Locator::contains('//label[@for="name"]', 'enter a name'));
+
+
+    }
+
     public function testHumanReadableString()
     {
        $this->assertEquals("'string selector'", Locator::humanReadableString("string selector"));